Реализованы itoa, обновлён тест, заголовок, скрипт для сборки

+++ include/e1_str.h
@@ -18,9 +18,9 @@ typedef struct {
 str_t readstr();
 void reverse(str_t *str);
 str_t itoa(int num, int base);
-str_t ltoa(long num, int base);
+str_t ltoa(long long num, int base);
 str_t utoa(unsigned int num, int base);
-str_t ultoa(unsigned long num, int base);
+str_t ultoa(unsigned long long num, int base);
 char* str_to_cstr(str_t *str);
 str_t cstr_to_str(char *str);
